Add rule to not audit when WLC accesses incremental_prop and proc
While full scan in WLC, if an app invokes incfs filesystem, it causes
an avc denial denial for accessing incremental_prop and proc
filesystem. Since, WLC does not directly access the incremental
property, and that doesn't affects the classification of the apk,
add a rule to not audit when wlc reads the incremental_prop
and proc_filesystems.

Change-Id: Ic89327a7c048b760c24a1fc272dd2ee848689699